Auth/users/associations

+++ src/routes/rsvpRouter.js
@@ -1,27 +1,30 @@
 const express = require("express");
 const QRcode = require("qrcode");
 const mongoose = require("mongoose");
+const auth = require("../middleware/auth");
 const RSVP = require("../models/RSVP");
-const acceptEmail = require("../email/acceptEmail");
+const {acceptEmail} = require("../email/Email");
 
 const rsvpRouter = new express.Router();
 
 rsvpRouter.get("/", (req, res) => {
-    res.redirect(`http://127.0.0.1:1337`);
+    res.status(200).redirect(`http://127.0.0.1:1337`);
 });
 
-rsvpRouter.post("/rsvp", async (req, res) => {
+rsvpRouter.post("/rsvp", auth, async (req, res) => {
     const id = new mongoose.Types.ObjectId();
     const qr = await QRcode.toDataURL(`localhost:3000/rsvp/${id}`); //todo
 
     const rsvp = new RSVP({
         ...req.body,
         qr,
-        id
+        id,
+        owner:req.user._id
     });
 
     try {
         await rsvp.save();
+        //creationEmail() //todo
         res.status(201).send(rsvp);
     } catch (err) {
         res.status(400).send(err);
@@ -38,8 +41,7 @@ rsvpRouter.get("/rsvp/:id", async (req, res) => {
     }
 });
 
-//Protected
-rsvpRouter.patch("/rsvp/:id", async (req, res) => {
+rsvpRouter.patch("/rsvp/:id", auth, async (req, res) => {
     const updates = Object.keys(req.body);
     const allowedUpdates = [
         "location",
@@ -56,7 +58,7 @@ rsvpRouter.patch("/rsvp/:id", async (req, res) => {
     if (!valid) return res.status(400).send({ error: "Invalid Updates" });
 
     try {
-        const rsvp = await RSVP.findOne({ id: req.params.id });
+        const rsvp = await RSVP.findOne({ id: req.params.id, owner:req.user._id });
         if (!rsvp) return res.status(404).send();
 
         updates.forEach((update) => rsvp[update] = req.body[update]);
@@ -69,15 +71,46 @@ rsvpRouter.patch("/rsvp/:id", async (req, res) => {
 
 });
 
-//Protected
-rsvpRouter.delete("/rsvp/:id", async (req, res) => {
+rsvpRouter.delete("/rsvp/:id", auth, async (req, res) => {
     try {
-        const rsvp = await RSVP.findOneAndDelete({ id: req.params.id });
+        const rsvp = await RSVP.findOneAndDelete({ id: req.params.id, owner: req.user._id });
         if (!rsvp) return res.status(404).send();
 
         res.status(200).send();
     } catch (err) {
         res.status(400).send(err);
+    }
+});
+
+rsvpRouter.get("/rsvps", auth, async (req, res) => {
+    // const match = {};
+    // if (req.query.completed) {
+    //     match.completed = req.query.completed === "true";
+    // }
+
+    // const sort = {};
+    // if (req.query.sortBy) {
+    //     const parts = req.query.sortBy.split(":");
+    //     sort[parts[0]] = parts[1] === "desc" ? -1 : 1;
+    // }
+
+    // {
+    //     path: "rsvps",
+    //     match,
+    //     options:{
+    //         limit: parseInt(req.query.limit),
+    //         skip: parseInt(req.query.skip),
+    //         sort
+    //     }
+    // }
+
+    try {
+        await req.user.populate({ //todo
+            path: "rsvps",
+        }).execPopulate()
+        res.status(200).send(req.user.rsvps);
+    } catch (err) {
+        res.status(500).send(err);
     }
 });
